# Avoid using join for StockLevel transaction.