--- a/src/com/oltpbenchmark/benchmarks/tpcc/TPCCWorker.java
+++ b/src/com/oltpbenchmark/benchmarks/tpcc/TPCCWorker.java
@@ -105,5 +105,9 @@ public class TPCCWorker extends Worker<TPCCBenchmark> {
     TPCCProcedure proc = (TPCCProcedure) this.getProcedure(
         this.transactionTypes.getType("NewOrder").getProcedureClass());
     proc.test(conn, this);
+
+    proc = (TPCCProcedure) this.getProcedure(
+        this.transactionTypes.getType("StockLevel").getProcedureClass());
+    proc.test(conn, this);
   }
 }