@robertsami 2021-02-19
lets close this in favor of https://github.com/yugabyte/tpcc/pull/75